% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_cloud.R
\docType{data}
\name{example_cloud}
\alias{example_cloud}
\title{example cloud}
\format{
\subsection{\code{example_cloud}}{

An example of a simulated cloud raster
}
}
\usage{
example_cloud
}
\description{
this is created by the file example_cloud.R in raw data
}
\keyword{datasets}
