% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_cloud_matched.R
\docType{data}
\name{example_cloud_matched}
\alias{example_cloud_matched}
\title{example cloud}
\format{
\subsection{\code{example_cloud_matched}}{

An example of a simulated cloud raster that has been matched to the raster
}
}
\usage{
example_cloud_matched
}
\description{
this is created by the file example_cloud.R in raw data
}
\keyword{datasets}
