% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_raster.R
\docType{data}
\name{example_raster}
\alias{example_raster}
\title{example raster}
\format{
\subsection{\code{example_raster}}{

An example raster object
}
}
\usage{
example_raster
}
\description{
this is created by the file example_raster.R in raw data
}
\keyword{datasets}
