% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cloud.R
\name{filter_cloud}
\alias{filter_cloud}
\title{filter_cloud}
\usage{
filter_cloud(cloud, type = "speckled")
}
\arguments{
\item{cloud}{a cloud raster}

\item{type}{a choice between speckled and large.}
}
\value{
cloud raster with NAs.
}
\description{
Add NAs to pixels around zero. These NAs then become the clouds
}
\examples{
library(nimbus)
library(raster)
cloud <- simulate_cloud(phi = 0.1, seed = 2023)
plot(cloud)
speckled_cloud_raster <- filter_cloud(cloud)
large_cloud_raster <- filter_cloud(cloud, type = "large")
plot(speckled_cloud_raster, colNA = "blue")
plot(large_cloud_raster, colNA = "blue")
}
