% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_cloud.R
\docType{data}
\name{large_cloud}
\alias{large_cloud}
\title{large cloud}
\format{
\subsection{\code{large_cloud}}{

An example of a large cloud raster
}
}
\usage{
large_cloud
}
\description{
large cloud
}
\keyword{datasets}
