% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_cloud.R
\name{simulate_cloud}
\alias{simulate_cloud}
\title{simulate cloud}
\usage{
simulate_cloud(cloud_size = 70, phi, seed = NULL, form = "smooth")
}
\arguments{
\item{cloud_size}{the size of the grid along each axis.}

\item{phi}{parameter to control the structure of the cloud}

\item{seed}{a seed to pass to ensure reproducibility}

\item{form}{type of cloud, either smooth or pixellated}
}
\value{
a raster object of simulated clouds
}
\description{
Simulates a cloud raster
}
\examples{
library(raster)
cloud <- simulate_cloud(cloud_size = 10, phi = 0.1)
plot(cloud)
}
