% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speckled_cloud.R
\docType{data}
\name{speckled_cloud}
\alias{speckled_cloud}
\title{speckled cloud}
\format{
\subsection{\code{speckled_cloud}}{

An example of a speckled cloud raster
}
}
\usage{
speckled_cloud
}
\description{
this is created by the file example_cloud.R in raw data
}
\keyword{datasets}
