% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cloud.R
\name{add_cloud}
\alias{add_cloud}
\title{add_cloud}
\usage{
add_cloud(raster, cloud)
}
\arguments{
\item{raster}{a raster object}

\item{cloud}{the cloud object to add}
}
\value{
raster object with clouds
}
\description{
add_cloud
}
\examples{
library(nimbus)
library(raster)
plot(example_raster)
cloud <- example_raster
cloud[cloud < 500] <- NA
example_raster |> add_cloud(cloud) |> plot()
}
