% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_rasters.R
\name{match_rasters}
\alias{match_rasters}
\title{match_rasters}
\usage{
match_rasters(raster, cloud, trace = FALSE)
}
\arguments{
\item{raster}{original raster to match to.}

\item{cloud}{simulated cloud}

\item{trace}{will show some feedback if needed.}
}
\value{
matched cloud raster
}
\description{
This function matches the simulated cloud so that it is on the same
extent and projection as the original raster
}
\examples{
library(raster)
plot(example_cloud)
example_cloud <- match_rasters(example_raster, example_cloud, trace = TRUE)
plot(example_cloud)
}
